+++ docs/changelog.md
@@ -18,6 +18,7 @@ This version is not fully backwords-compatible. Check out [the migration guide](
 - 3e672ee `journeys()`/`journeyLeg()`: `stopover.station` → `stopover.stop` 💥
 - 021ae45: `journeys()`/`journeyLeg()`: leg stopovers: parse & expose platforms
 - 85e0bdf `journeys()`: `startWithWalking` option with default `true` ✨
+- 2e6aefe journey leg, departure, movement: `journeyId` -> `tripId` 💥
 
 ## `2.7.0`